requirement.txt
streamlit
langchain
openai
anthropic
google-generativeai
moviepy
pillow
transformers
langchain-community